Refactor ShapefileWriter and cache DriverInfo in OGR utilities
