+++ hoot-core/src/main/cpp/hoot/core/io/DataConverter.cpp
@@ -32,10 +32,10 @@
 #include <hoot/core/io/ElementCacheLRU.h>
 #include <hoot/core/io/ElementStreamer.h>
 #include <hoot/core/io/ElementTranslatorThread.h>
-#include <hoot/core/io/OsmGeoJsonWriter.h>
 #include <hoot/core/io/IoUtils.h>
 #include <hoot/core/io/OgrWriter.h>
 #include <hoot/core/io/OgrWriterThread.h>
+#include <hoot/core/io/OsmGeoJsonWriter.h>
 #include <hoot/core/io/OsmMapReader.h>
 #include <hoot/core/io/OsmMapReaderFactory.h>
 #include <hoot/core/io/OsmMapWriterFactory.h>
@@ -120,8 +120,7 @@ void DataConverter::convert(const QStringList& inputs, const QString& output)
     "Converting ..." + FileUtils::toLogFormat(inputs, _printLengthMax) + " to ..." +
     FileUtils::toLogFormat(output, _printLengthMax) + "...");
 
-  const bool isStreamable =
-    IoUtils::areValidStreamingOps(_convertOps) && IoUtils::areStreamableIo(inputs, output);
+  const bool isStreamable = IoUtils::areValidStreamingOps(_convertOps) && IoUtils::areStreamableIo(inputs, output);
   LOG_VARD(isStreamable);
   //  Running the translator in a separate thread from the writer is an option for OGR, but the I/O
   //  formats must be streamable.
@@ -325,15 +324,13 @@ void DataConverter::_exportToShapeWithCols(const QString& output, const QStringL
   QElapsedTimer timer;
   timer.start();
 
-  std::shared_ptr<OsmMapWriter> writer =
-    OsmMapWriterFactory::createWriter(output);
-  std::shared_ptr<ShapefileWriter> shapeFileWriter =
-    std::dynamic_pointer_cast<ShapefileWriter>(writer);
+  std::shared_ptr<OsmMapWriter> writer = OsmMapWriterFactory::createWriter(output);
+  std::shared_ptr<ShapefileWriter> shapeFileWriter = std::dynamic_pointer_cast<ShapefileWriter>(writer);
   //  Currently there is only one shape file writer, and this is it.
   assert(shapeFileWriter.get());
   shapeFileWriter->setColumns(cols);
   shapeFileWriter->open(output);
-  shapeFileWriter->write(map, output);
+  shapeFileWriter->write(map);
 
   LOG_INFO(
     "Wrote " << StringUtils::formatLargeNumber(map->getElementCount()) <<
@@ -343,8 +340,7 @@ void DataConverter::_exportToShapeWithCols(const QString& output, const QStringL
 void DataConverter::_fillElementCacheMT(const QString& inputUrl, ElementCachePtr cachePtr, QQueue<ElementPtr>& workQ) const
 {
   //  Setup reader
-  std::shared_ptr<OsmMapReader> reader =
-    OsmMapReaderFactory::createReader(inputUrl);
+  std::shared_ptr<OsmMapReader> reader = OsmMapReaderFactory::createReader(inputUrl);
   reader->open(inputUrl);
   std::shared_ptr<ElementInputStream> streamReader = std::dynamic_pointer_cast<ElementInputStream>(reader);
 
@@ -469,9 +465,7 @@ void DataConverter::_setToOgrOptions(const QString& output)
   //  warn callers that the opposite direction they specified won't be used.
   if (_translationDirection == "toosm")
   {
-    LOG_INFO(
-      "Ignoring specified schema.translation.direction=toosm and using toogr to write to " <<
-      "OGR output...");
+    LOG_INFO("Ignoring specified schema.translation.direction=toosm and using toogr to write to OGR output...");
   }
 
   //  Set a config option so the translation script knows what the output format is. For this,
@@ -495,8 +489,7 @@ void DataConverter::_handleNonOgrOutputTranslationOpts()
   assert(!_shapeFileColumnsSpecified());
 
   //  For non OGR conversions, the translation must be passed in as an operator.
-  if (!_convertOps.contains(SchemaTranslationOp::className()) &&
-      !_convertOps.contains(SchemaTranslationVisitor::className()))
+  if (!_convertOps.contains(SchemaTranslationOp::className()) && !_convertOps.contains(SchemaTranslationVisitor::className()))
   {
     //  If a translation script was specified but not the translation op, we'll add auto add the op
     //  as the first conversion operation. If the caller wants the translation done after some
