+++ hoot-core/src/main/cpp/hoot/core/io/OgrMultifileWriter.cpp
@@ -1,0 +1,466 @@
+/*
+ * This file is part of Hootenanny.
+ *
+ * Hootenanny is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ * --------------------------------------------------------------------
+ *
+ * The following copyright notices are generated automatically. If you
+ * have a new notice to add, please use the format:
+ * " * @copyright Copyright ..."
+ * This will properly maintain the copyright information. Maxar
+ * copyrights will be updated automatically.
+ *
+ * @copyright Copyright (C) 2023 Maxar (http://www.maxar.com/)
+ */
+
+#include "OgrMultifileWriter.h"
+
+// GDAL
+#include <ogr_geometry.h>
+
+// GEOS
+#include <geos/geom/LineString.h>
+#include <geos/geom/Polygon.h>
+
+// Hoot
+#include <hoot/core/criterion/AreaCriterion.h>
+#include <hoot/core/elements/OsmMap.h>
+#include <hoot/core/elements/Relation.h>
+#include <hoot/core/elements/Way.h>
+#include <hoot/core/geometry/ElementToGeometryConverter.h>
+#include <hoot/core/io/OgrOptions.h>
+#include <hoot/core/schema/MetadataTags.h>
+#include <hoot/core/util/ConfigOptions.h>
+#include <hoot/core/util/FileUtils.h>
+#include <hoot/core/visitors/ElementConstOsmMapVisitor.h>
+
+using namespace geos::geom;
+
+namespace hoot
+{
+
+class ColumnVisitor : public ElementConstOsmMapVisitor
+{
+public:
+
+  ColumnVisitor(ElementType type) : _type(type) { }
+  ~ColumnVisitor() override = default;
+
+  void visit(const std::shared_ptr<const Element>& e) override
+  {
+    if (e->getElementType() == _type || _type == ElementType::Unknown)
+    {
+      const Tags& tags = e->getTags();
+      for (auto it = tags.begin(); it != tags.end(); ++it)
+        _keys[it.key()]++;
+    }
+  }
+
+  QMultiMap<int, QString> getBackwardKeys()
+  {
+    QMultiMap<int, QString> result;
+    for (auto it = _keys.begin(); it != _keys.end(); ++it)
+      result.insertMulti(it.value(), it.key());
+    return result;
+  }
+
+  QString getDescription() const override { return ""; }
+  QString getName() const override { return ""; }
+  QString getClassName() const override { return ""; }
+
+private:
+
+  ElementType _type;
+  QHash<QString, int> _keys;
+};
+
+OgrMultifileWriter::OgrMultifileWriter()
+  : _includeCircularError(ConfigOptions().getWriterIncludeCircularErrorTags()),
+    _circularErrorIndex(-1)
+{
+}
+
+QStringList OgrMultifileWriter::getColumns(ConstOsmMapPtr map, ElementType type) const
+{
+  if (!_columns.empty())
+    return _columns;
+  else
+  {
+    QStringList result;
+    ColumnVisitor v(type);
+    map->visitRo(v);
+    QMultiMap<int, QString> backwardKeys = v.getBackwardKeys();
+
+    QMapIterator<int, QString> i(backwardKeys);
+    i.toBack();
+    while (i.hasPrevious() && result.size() < 20)
+    {
+      i.previous();
+      result << i.value();
+    }
+
+    return result;
+  }
+}
+
+void OgrMultifileWriter::open(const QString& url)
+{
+  if (QDir(url).exists() == false)
+  {
+    if (FileUtils::makeDir(url) == false)
+      throw HootException("Error creating directory for writing: " + url);
+  }
+  _outputDir = QDir(url);
+}
+
+void OgrMultifileWriter::setConfiguration(const Settings& conf)
+{
+  _includeCircularError = ConfigOptions(conf).getWriterIncludeCircularErrorTags();
+}
+
+void OgrMultifileWriter::write(const ConstOsmMapPtr& map)
+{
+  writeLines(map, _outputDir.absoluteFilePath("Lines" + _getFileExtension()));
+  writePoints(map, _outputDir.absoluteFilePath("Points" + _getFileExtension()));
+  writePolygons(map, _outputDir.absoluteFilePath("Polygons" + _getFileExtension()));
+}
+
+void OgrMultifileWriter::writeLines(const ConstOsmMapPtr& map, const QString& path)
+{
+  GDALAllRegister();
+
+  _removeMultifile(path);
+
+  const char* pszDriverName = "ESRI Shapefile";
+  GDALDriver* poDriver = GetGDALDriverManager()->GetDriverByName(pszDriverName);
+  if (poDriver == nullptr)
+    throw HootException(QString("%1 driver not available.").arg(pszDriverName));
+
+  GDALDataset* poDS = poDriver->Create(path.toLatin1(), 0, 0, 0, GDT_Unknown, nullptr);
+  if (poDS == nullptr)
+    throw HootException(QString("Data source creation failed. %1").arg(path));
+
+  OGRLayer* poLayer;
+
+  OgrOptions options;
+  options["ENCODING"] = "UTF-8";
+
+  QString layerName;
+  layerName = QFileInfo(path).baseName();
+  poLayer = poDS->CreateLayer(layerName.toLatin1(),
+                              map->getProjection().get(), wkbLineString,
+                              options.getCrypticOptions());
+  if (poLayer == nullptr)
+    throw HootException(QString("Layer creation failed. %1").arg(path));
+
+  QStringList shpColumns;
+  QStringList columns = getColumns(map, ElementType::Way);
+
+  for (int i = 0; i < columns.size(); i++)
+  {
+    OGRFieldDefn oField(columns[i].toLatin1(), OFTString);
+    oField.SetWidth(64);
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(columns[i]));
+
+    shpColumns.append(poLayer->GetLayerDefn()->GetFieldDefn(i)->GetNameRef());
+  }
+
+  if (_includeCircularError)
+  {
+    OGRFieldDefn oField("error_circ", OFTReal);
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(MetadataTags::ErrorCircular()));
+    _circularErrorIndex = poLayer->GetLayerDefn()->GetFieldCount() - 1;
+  }
+
+  const WayMap& ways = map->getWays();
+  for (auto it = ways.begin(); it != ways.end(); ++it)
+  {
+    WayPtr way = it->second;
+
+    if (AreaCriterion().isSatisfied(way) == false)
+    {
+      OGRFeature* poFeature = OGRFeature::CreateFeature(poLayer->GetLayerDefn());
+      // set all the column values.
+      for (int i = 0; i < columns.size(); i++)
+      {
+        if (way->getTags().contains(columns[i]))
+        {
+          QByteArray c = shpColumns[i].toLatin1();
+          QByteArray v = way->getTags()[columns[i]].toUtf8();
+          poFeature->SetField(c.constData(), v.constData());
+        }
+      }
+
+      if (_includeCircularError)
+        poFeature->SetField(_circularErrorIndex, way->getCircularError());
+
+      // convert the geometry.
+      std::string wkt = ElementToGeometryConverter(map).convertToLineString(way)->toString();
+      const char* t = wkt.data();
+      OGRGeometry* geom;
+      if (OGRGeometryFactory::createFromWkt(&t, poLayer->GetSpatialRef(), &geom) != OGRERR_NONE)
+        throw HootException(QString("Error parsing WKT (%1)").arg(QString::fromStdString(wkt)));
+
+      if (poFeature->SetGeometryDirectly(geom) != OGRERR_NONE)
+        throw HootException(QString("Error setting geometry"));
+
+      if (poLayer->CreateFeature(poFeature) != OGRERR_NONE)
+        throw HootException(QString("Error creating feature"));
+
+      OGRFeature::DestroyFeature(poFeature);
+    }
+  }
+
+  GDALClose(poDS);
+}
+
+void OgrMultifileWriter::writePoints(const ConstOsmMapPtr& map, const QString& path)
+{
+  GDALAllRegister();
+
+  _removeMultifile(path);
+
+  const char* pszDriverName = "ESRI Shapefile";
+  GDALDriver* poDriver = GetGDALDriverManager()->GetDriverByName(pszDriverName);
+  if (poDriver == nullptr)
+    throw HootException(QString("%1 driver not available.").arg(pszDriverName));
+
+  GDALDataset* poDS = poDriver->Create(path.toLatin1(), 0, 0, 0, GDT_Unknown, nullptr);
+  if (poDS == nullptr)
+    throw HootException(QString("Data source creation failed. %1").arg(path));
+
+  OgrOptions options;
+  options["ENCODING"] = "UTF-8";
+
+  OGRLayer* poLayer;
+
+  QString layerName;
+  layerName = QFileInfo(path).baseName();
+  poLayer = poDS->CreateLayer(layerName.toLatin1(), map->getProjection().get(),
+                              wkbPoint, options.getCrypticOptions());
+  if (poLayer == nullptr)
+    throw HootException(QString("Layer creation failed. %1").arg(path));
+
+  QStringList shpColumns;
+  QStringList columns = getColumns(map, ElementType::Node);
+
+  for (int i = 0; i < columns.size(); i++)
+  {
+    OGRFieldDefn oField(columns[i].toLatin1(), OFTString);
+
+    oField.SetWidth(64);
+
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(columns[i]));
+
+    shpColumns.append(poLayer->GetLayerDefn()->GetFieldDefn(i)->GetNameRef());
+  }
+
+  if (_includeCircularError)
+  {
+    OGRFieldDefn oField("error_circ", OFTReal);
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(MetadataTags::ErrorCircular()));
+    _circularErrorIndex = poLayer->GetLayerDefn()->GetFieldCount() - 1;
+  }
+
+  const NodeMap& nodes = map->getNodes();
+  for (auto it = nodes.begin(); it != nodes.end(); ++it)
+  {
+    const NodePtr& node = it->second;
+
+    if (node->getTags().getNonDebugCount() > 0)
+    {
+      OGRFeature* poFeature = OGRFeature::CreateFeature(poLayer->GetLayerDefn());
+      // set all the column values.
+      for (int i = 0; i < columns.size(); i++)
+      {
+        if (node->getTags().contains(columns[i]))
+        {
+          QByteArray c = shpColumns[i].toLatin1();
+          QByteArray v = node->getTags()[columns[i]].toUtf8();
+          poFeature->SetField(c.constData(), v.constData());
+        }
+      }
+      if (_includeCircularError)
+        poFeature->SetField(_circularErrorIndex, node->getCircularError());
+
+      // convert the geometry.
+      std::shared_ptr<OGRGeometry> geom = std::make_shared<OGRPoint>(node->getX(), node->getY());
+
+      if (poFeature->SetGeometry(geom.get()) != OGRERR_NONE)
+        throw HootException(QString("Error setting geometry"));
+
+      if (poLayer->CreateFeature(poFeature) != OGRERR_NONE)
+        throw HootException(QString("Error creating feature"));
+
+      OGRFeature::DestroyFeature(poFeature);
+    }
+  }
+
+  GDALClose(poDS);
+}
+
+void OgrMultifileWriter::writePolygons(const ConstOsmMapPtr& map, const QString& path)
+{
+  GDALAllRegister();
+
+  _removeMultifile(path);
+
+  const char* pszDriverName = "ESRI Shapefile";
+  GDALDriver* poDriver = GetGDALDriverManager()->GetDriverByName(pszDriverName);
+  if (poDriver == nullptr)
+    throw HootException(QString("%1 driver not available.").arg(pszDriverName));
+
+  GDALDataset* poDS = poDriver->Create(path.toLatin1(), 0, 0, 0, GDT_Unknown, nullptr);
+  if (poDS == nullptr)
+    throw HootException(QString("Data source creation failed. %1").arg(path));
+
+  OgrOptions options;
+  options["ENCODING"] = "UTF-8";
+
+  OGRLayer *poLayer;
+
+  QString layerName;
+  layerName = QFileInfo(path).baseName();
+  poLayer = poDS->CreateLayer(layerName.toLatin1(),
+                              map->getProjection().get(), wkbMultiPolygon,
+                              options.getCrypticOptions());
+  if (poLayer == nullptr)
+    throw HootException(QString("Layer creation failed. %1").arg(path));
+
+  QStringList shpColumns;
+  QStringList columns = getColumns(map, ElementType::Unknown);
+
+  for (int i = 0; i < columns.size(); i++)
+  {
+    OGRFieldDefn oField(columns[i].toLatin1(), OFTString);
+    oField.SetWidth(64);
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(columns[i]));
+
+    shpColumns.append(poLayer->GetLayerDefn()->GetFieldDefn(i)->GetNameRef());
+  }
+
+  if (_includeCircularError)
+  {
+    OGRFieldDefn oField("error_circ", OFTReal);
+    if (poLayer->CreateField(&oField) != OGRERR_NONE)
+      throw HootException(QString("Error creating field (%1).").arg(MetadataTags::ErrorCircular()));
+    _circularErrorIndex = poLayer->GetLayerDefn()->GetFieldCount() - 1;
+  }
+
+  const WayMap& ways = map->getWays();
+  for (auto it = ways.begin(); it != ways.end(); ++it)
+  {
+    WayPtr way = it->second;
+
+    if (AreaCriterion().isSatisfied(way))
+      _writeWayPolygon(map, way, poLayer, columns, shpColumns);
+  }
+
+  const RelationMap& relations = map->getRelations();
+  for (auto it = relations.begin(); it != relations.end(); ++it)
+  {
+    RelationPtr relation = it->second;
+
+    if (relation->isMultiPolygon())
+      _writeRelationPolygon(map, relation, poLayer, columns, shpColumns);
+  }
+
+  GDALClose(poDS);
+}
+
+void OgrMultifileWriter::_writeRelationPolygon(const ConstOsmMapPtr& map, const RelationPtr& relation,
+                                            OGRLayer* poLayer, const QStringList& columns,
+                                            const QStringList& shpColumns) const
+{
+  OGRFeature* poFeature = OGRFeature::CreateFeature(poLayer->GetLayerDefn());
+  // set all the column values.
+  for (int i = 0; i < columns.size(); i++)
+  {
+    if (relation->getTags().contains(columns[i]))
+    {
+      QByteArray c = shpColumns[i].toLatin1();
+      QByteArray v = relation->getTags()[columns[i]].toUtf8();
+      poFeature->SetField(c.constData(), v.constData());
+    }
+  }
+
+  if (_includeCircularError)
+    poFeature->SetField(_circularErrorIndex, relation->getCircularError());
+
+  // convert the geometry.
+  const ConstRelationPtr& r = relation;
+  std::shared_ptr<geos::geom::Geometry> geometry = ElementToGeometryConverter(map).convertToGeometry(r);
+  if (!geometry.get() || geometry->isEmpty())
+    throw HootException(QString("Error converting geometry"));
+
+  std::string wkt = geometry->toString();
+  const char* t = wkt.data();
+  OGRGeometry* geom;
+  if (OGRGeometryFactory::createFromWkt(&t, poLayer->GetSpatialRef(), &geom) != OGRERR_NONE)
+    throw HootException(QString("Error parsing WKT (%1)").arg(QString::fromStdString(wkt)));
+
+  if (poFeature->SetGeometryDirectly(geom) != OGRERR_NONE)
+    throw HootException(QString("Error setting geometry"));
+
+  if (poLayer->CreateFeature(poFeature) != OGRERR_NONE)
+    throw HootException(QString("Error creating feature"));
+
+  OGRFeature::DestroyFeature(poFeature);
+}
+
+void OgrMultifileWriter::_writeWayPolygon(const ConstOsmMapPtr& map, const WayPtr& way, OGRLayer* poLayer,
+                                       const QStringList& columns, const QStringList &shpColumns) const
+{
+  OGRFeature* poFeature = OGRFeature::CreateFeature(poLayer->GetLayerDefn());
+  // set all the column values.
+  for (int i = 0; i < columns.size(); i++)
+  {
+    if (way->getTags().contains(columns[i]))
+    {
+      QByteArray c = shpColumns[i].toLatin1();
+      QByteArray v = way->getTags()[columns[i]].toUtf8();
+      poFeature->SetField(c.constData(), v.constData());
+    }
+  }
+
+  if (_includeCircularError)
+    poFeature->SetField(_circularErrorIndex, way->getCircularError());
+
+  // convert the geometry.
+  std::shared_ptr<Geometry> p = ElementToGeometryConverter(map).convertToGeometry(way);
+  if (p->getGeometryTypeId() != GEOS_POLYGON)
+    throw InternalErrorException("Expected a polygon geometry, but got a: " + toString(p->getGeometryType()));
+
+  std::string wkt = p->toString();
+  const char* t = wkt.data();
+  OGRGeometry* geom;
+  if (OGRGeometryFactory::createFromWkt(&t, poLayer->GetSpatialRef(), &geom) != OGRERR_NONE)
+    throw HootException(QString("Error parsing WKT (%1)").arg(QString::fromStdString(wkt)));
+
+  if (poFeature->SetGeometryDirectly(geom) != OGRERR_NONE)
+    throw HootException(QString("Error setting geometry"));
+
+  if (poLayer->CreateFeature(poFeature) != OGRERR_NONE)
+    throw HootException(QString("Error creating feature"));
+
+  OGRFeature::DestroyFeature(poFeature);
+}
+
+}
